+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @builder.io/mitosis-cli
 
+## 0.12.0
+
+### Minor Changes
+
+- 2bd0070: feat: Builder Text node is no longer stripped away by default when converting to Mitosis
+
+### Patch Changes
+
+- Updated dependencies [2bd0070]
+  - @builder.io/mitosis@0.12.0
+
+## 0.11.6
+
+### Patch Changes
+
+- 314b396: Builder: improve accuracy of invalid binding detection
+- f442a8c: fix: support for builder custom component with a colon in their name
+- Updated dependencies [314b396]
+- Updated dependencies [f442a8c]
+  - @builder.io/mitosis@0.11.6
+
 ## 0.11.5
 
 ### Patch Changes